+++ .pre-commit-config.yaml
@@ -19,6 +19,7 @@ repos:
     rev: v0.3.7
     hooks:
       - id: ruff
+        args: [--fix]
 
   - repo: https://github.com/igorshubovych/markdownlint-cli
     rev: v0.33.0
