+++ tests/torch/fx/performance_check/requirements.txt
@@ -1,0 +1,2 @@
+diffusers
+ultralytics
